stpms/backend/requirements.txt
fastapi[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
redis
google-generativeai
langchain
langchain-google-genai
pydantic
pydantic-settings
